refactor(security): Added a centralize server TLS policy to reuse the TLS 1.2 minimum configuration across UTxORPC and Bark.
