+++ bark/bark.go
@@ -32,6 +32,7 @@ import (
 	archiveconnect "github.com/blinklabs-io/bark/proto/v1alpha1/archive/archivev1alpha1connect"
 	"github.com/blinklabs-io/dingo/database"
 	"github.com/blinklabs-io/dingo/internal/httpcors"
+	"github.com/blinklabs-io/dingo/internal/tlsutil"
 )
 
 type Bark struct {
@@ -193,18 +194,6 @@ func unencryptedHTTP2Protocols() *http.Protocols {
 	return protocols
 }
 
-func configureServerTLS(server *http.Server) {
-	if server.TLSConfig == nil {
-		server.TLSConfig = &tls.Config{
-			MinVersion: tls.VersionTLS12,
-		}
-		return
-	}
-	if server.TLSConfig.MinVersion < tls.VersionTLS12 {
-		server.TLSConfig.MinVersion = tls.VersionTLS12
-	}
-}
-
 // startServer starts the HTTP server with deterministic error
 // detection. It validates TLS configuration, binds the listening
 // socket and pre-loads any TLS keypair synchronously so port and
@@ -229,7 +218,7 @@ func (b *Bark) startServer(server *http.Server) error {
 				serverType, err,
 			)
 		}
-		configureServerTLS(server)
+		server.TLSConfig = tlsutil.ServerConfig(server.TLSConfig)
 	}
 	ln, err := net.Listen("tcp", server.Addr)
 	if err != nil {
